requirements.txt
PyQt6==6.6.1
JPype1==1.5.0
py4j==0.10.9.7
